requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
httpx>=0.27.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
itsdangerous>=2.2.0
passlib>=1.7.4
bcrypt>=4.1.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
gunicorn>=21.2.0
